# docs/openapi.yaml
openapi: 3.0.3
info:
  title: sandboxd /v1 API
  version: "0.3.0"
  description: >
    The public, tenant-scoped API for sandboxd. This is the only surface a
    console or external integration should use — the internal /sandbox and
    /wake endpoints are not part of this contract. Scoped by the API token
    (Bearer); when auth is disabled (the single-user default) the whole
    surface is reachable without a token.
servers:
  - url: http://127.0.0.1:9090
    description: Default loopback bind (docker-compose maps SANDBOXD_API_BIND here)
security:
  - bearerAuth: []
  - {} # auth disabled by default (single-user)
...
paths:
...
  /v1/sandboxes/{id}/tasks/{taskId}:
    parameters:
      - $ref: "#/components/parameters/Id"
      - in: path
        name: taskId
        required: true
        schema: { type: string }
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task's canonical result
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskResult" }
        "404": { $ref: "#/components/responses/Error" }
...
  /v1/sandboxes/{id}/tasks/{taskId}/cancel:
    parameters:
      - $ref: "#/components/parameters/Id"
      - in: path
        name: taskId
        required: true
        schema: { type: string }
    post:
      operationId: cancelTask
      tags: [tasks]
      summary: Cancel a running task
      responses:
        "200": { description: Cancelled }
  /v1/sandboxes/{id}/tasks/{taskId}/revert:
    parameters:
      - $ref: "#/components/parameters/Id"
      - in: path
        name: taskId
        required: true
        schema: { type: string }
    post:
      operationId: revertTask
      tags: [tasks]
      summary: Revert the workspace to a task's pre-task checkpoint
      description: >
        Restores the working files to how they were before the task ran (undoing
        that task and everything after it), without moving the branch/HEAD.
        Requires a running sandbox; the task must have a checkpoint.
      responses:
        "200": { description: Reverted }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            retryable: { type: boolean }
...
    TaskResult:
      type: object
      properties:
        id: { type: string }
        sandbox_id: { type: string }
        status: { type: string, enum: [running, succeeded, failed, cancelled] }
        failure_reason: { type: string }
        error_message: { type: string }
        files_changed: { type: array, items: { type: string } }
        build_ok: { type: boolean, description: "Backward-compat; true only when build_status=passed." }
        build_status: { type: string, enum: [passed, failed, skipped] }
        build_error_message: { type: string }
        preview_status_after: { type: string }
        preview_ok: { type: boolean, description: "Web apps only; omitted for worker-only." }
        app_healthy: { type: boolean }
        duration_ms: { type: integer }
...